Update: freq_compile

+++ freq/freq.c
@@ -3,8 +3,8 @@
 void
 main(){
 	struct timeval tv1, tv2;
-	long time,cycle_1_l,cycle_2_l,cycle_1_h,cycle_2_h;
-	long long cycle_1,cycle_2;
+	unsigned long time,cycle_1_l,cycle_2_l,cycle_1_h,cycle_2_h;
+	unsigned long cycle_1,cycle_2;
 	double cpufreq;
 	while(1){
 		gettimeofday(&tv1, NULL);
@@ -22,8 +22,8 @@ main(){
 			: "=a" (cycle_2_l), "=d" (cycle_2_h), "=r" (cycle_1_l), "=r" (cycle_1_h)
 		);
 		gettimeofday(&tv2, NULL);
-		cycle_1 = (cycle_1_h << 32) | (cycle_1_l & 0x0fffffff);
-		cycle_2 = (cycle_2_h << 32) | (cycle_2_l & 0x0fffffff);
+		cycle_1 = ((unsigned long)cycle_1_h << 32) | (cycle_1_l & 0x0fffffff);
+		cycle_2 = ((unsigned long)cycle_2_h << 32) | (cycle_2_l & 0x0fffffff);
 		time = (tv2.tv_sec - tv1.tv_sec) * 1000000 + (tv2.tv_usec - tv1.tv_usec);
 		cpufreq = (cycle_2 - cycle_1) * 1.0 / (time * 1.0);
 		printf("The cpu frequency is %.3f MHz, time = %d us, cycle = %lu,%d\n",cpufreq,time,cycle_2 - cycle_1,sizeof(cycle_1_h));
